+++ R/get_label.R
@@ -3,7 +3,7 @@
 #'
 #' @description This function returns the variable labels of labelled data.
 #'
-#' @seealso See package vignettes or \href{http://www.strengejacke.de/sjPlot/}{online documentation}
+#' @seealso See vignette \href{../doc/intro_sjmisc.html}{Labelled Data and the sjmisc-Package}
 #'            for more details; \code{\link{set_label}} to manually set variable labels or \code{\link{get_labels}}
 #'            to get value labels; \code{\link{var_labels}} to set multiple variable
 #'            labels at once.