+++ R/get_labels.R
@@ -3,7 +3,7 @@
 #'
 #' @description This function returns the value labels of labelled data.
 #'
-#' @seealso See \href{http://www.strengejacke.de/sjPlot/}{online documentation}
+#' @seealso See vignette \href{../doc/intro_sjmisc.html}{Labelled Data and the sjmisc-Package}
 #'            for more details; \code{\link{set_labels}} to manually set value
 #'            labels, \code{\link{get_label}} to get variable labels and
 #'            \code{\link{get_values}} to retrieve the values associated
@@ -69,10 +69,7 @@
 #'        Most functions of the \CRANpkg{sjPlot} package make use of value and variable
 #'        labels to automatically label axes, legend or title labels in plots
 #'        (\code{sjp.}-functions) respectively column or row headers in table
-#'        outputs (\code{sjt.}-functions). See
-#'        \href{http://www.strengejacke.de/sjPlot/datainit/}{this} and
-#'        \href{http://www.strengejacke.de/sjPlot/labelleddata/}{this}
-#'        online manual for more details.
+#'        outputs (\code{sjt.}-functions).
 #'
 #' @examples
 #' # import SPSS data set